+++ markup/add-review.html
@@ -96,34 +96,34 @@
           <div class="rating">
             <div class="rating__stars">
               <input class="rating__input" id="star-10" type="radio" name="rating" value="10" />
-              <label class="rating__label" for="star-10">Rating 10</label>
+              <label class="rating__label" for="star-10">RatingStars 10</label>
 
               <input class="rating__input" id="star-9" type="radio" name="rating" value="9" />
-              <label class="rating__label" for="star-9">Rating 9</label>
+              <label class="rating__label" for="star-9">RatingStars 9</label>
 
               <input class="rating__input" id="star-8" type="radio" name="rating" value="8" checked />
-              <label class="rating__label" for="star-8">Rating 8</label>
+              <label class="rating__label" for="star-8">RatingStars 8</label>
 
               <input class="rating__input" id="star-7" type="radio" name="rating" value="7" />
-              <label class="rating__label" for="star-7">Rating 7</label>
+              <label class="rating__label" for="star-7">RatingStars 7</label>
 
               <input class="rating__input" id="star-6" type="radio" name="rating" value="6" />
-              <label class="rating__label" for="star-6">Rating 6</label>
+              <label class="rating__label" for="star-6">RatingStars 6</label>
 
               <input class="rating__input" id="star-5" type="radio" name="rating" value="5" />
-              <label class="rating__label" for="star-5">Rating 5</label>
+              <label class="rating__label" for="star-5">RatingStars 5</label>
 
               <input class="rating__input" id="star-4" type="radio" name="rating" value="4" />
-              <label class="rating__label" for="star-4">Rating 4</label>
+              <label class="rating__label" for="star-4">RatingStars 4</label>
 
               <input class="rating__input" id="star-3" type="radio" name="rating" value="3" />
-              <label class="rating__label" for="star-3">Rating 3</label>
+              <label class="rating__label" for="star-3">RatingStars 3</label>
 
               <input class="rating__input" id="star-2" type="radio" name="rating" value="2" />
-              <label class="rating__label" for="star-2">Rating 2</label>
+              <label class="rating__label" for="star-2">RatingStars 2</label>
 
               <input class="rating__input" id="star-1" type="radio" name="rating" value="1" />
-              <label class="rating__label" for="star-1">Rating 1</label>
+              <label class="rating__label" for="star-1">RatingStars 1</label>
             </div>
           </div>
