+++ src/components/FilmsList.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import FilmCard from './FilmCard';
+import SmallFilmCard from './SmallFilmCard';
 import IFilm from "../types/IFilm";
 
 interface FilmListProps {
@@ -9,7 +9,7 @@ interface FilmListProps {
 const FilmsList: React.FC = (props: FilmListProps) => {
   return (
     <div className="catalog__films-list">
-      {props.films.map((film) => <FilmCard title={film.title} poster={film.poster} />)}
+      {props.films.map((film) => <SmallFilmCard title={film.title} poster={film.poster} />)}
     </div>
   )
 }
