+++ src/pages/film-page/FilmPage.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import Footer from "../../components/Footer";
 import Header from "../../components/Header";
 import MoreLikeThis from "../../components/MoreLikeThis";
+import FilmCardButtons from "../../components/FilmCardButtons";
 
 const FilmPage: React.FC = () => {
   return (
@@ -24,22 +25,7 @@ const FilmPage: React.FC = () => {
                 <span className="film-card__year">2014</span>
               </p>
 
-              <div className="film-card__buttons">
-                <button className="btn btn--play film-card__button" type="button">
-                  <svg viewBox="0 0 19 19" width="19" height="19">
-                    <use xlinkHref="#play-s"></use>
-                  </svg>
-                  <span>Play</span>
-                </button>
-                <button className="btn btn--list film-card__button" type="button">
-                  <svg viewBox="0 0 19 20" width="19" height="20">
-                    <use xlinkHref="#add"></use>
-                  </svg>
-                  <span>My list</span>
-                  <span className="film-card__count">9</span>
-                </button>
-                <a href="" className="btn film-card__button">Add review</a>
-              </div>
+              <FilmCardButtons />
             </div>
           </div>
         </div>
