+++ src/pages/main-page/MainPage.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import FilmsList from '../../components/FilmsList';
-import UserBlock from "../../components/UserBlock";
 import Footer from "../../components/Footer";
 import films from "../../utils/films";
-import Logo from "../../components/Logo";
 import CatalogGenreList from "../../components/CatalogGenreList";
+import FilmCard from "../../components/FilmCard";
 
 interface MainProps {
   title: string;
@@ -17,52 +16,8 @@ const MainPage: React.FC = (props: MainProps) => {
 
   return (
     <>
-      <section className="film-card">
-        <div className="film-card__bg">
-          <img src="markup/img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel"/>
-        </div>
+      <FilmCard title={title} genre={genre} date={date} />
 
-        <h1 className="visually-hidden">WTW</h1>
-
-        <header className="page-header film-card__head">
-          <Logo />
-          <UserBlock />
-
-        </header>
-
-        <div className="film-card__wrap">
-          <div className="film-card__info">
-            <div className="film-card__poster">
-              <img src="markup/img/the-grand-budapest-hotel-poster.jpg" alt="The Grand Budapest Hotel poster" width="218"
-                   height="327"/>
-            </div>
-
-            <div className="film-card__desc">
-              <h2 className="film-card__title">{title}</h2>
-              <p className="film-card__meta">
-                <span className="film-card__genre">{genre}</span>
-                <span className="film-card__year">{date}</span>
-              </p>
-
-              <div className="film-card__buttons">
-                <button className="btn btn--play film-card__button" type="button">
-                  <svg viewBox="0 0 19 19" width="19" height="19">
-                    <use xlinkHref="#play-s"></use>
-                  </svg>
-                  <span>Play</span>
-                </button>
-                <button className="btn btn--list film-card__button" type="button">
-                  <svg viewBox="0 0 19 20" width="19" height="20">
-                    <use xlinkHref="#add"></use>
-                  </svg>
-                  <span>My list</span>
-                  <span className="film-card__count">9</span>
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
       <div className="page-content">
         <section className="catalog">
           <h2 className="catalog__title visually-hidden">Catalog</h2>
